+++ testbed2d/src/demos/ballJoints.js
@@ -1,13 +1,11 @@
 export function initWorld(RAPIER, testbed) {
     let gravity = new RAPIER.Vector2(0.0, -9.81);
     let world = new RAPIER.World(gravity);
-    let bodies = new Array();
-    let colliders = new Array();
-    let joints = new Array();
+    let bodies = [];
 
     let rad = 0.4;
-    let numi = 70; // Num vertical nodes.
-    let numk = 70; // Num horizontal nodes.
+    let numi = 50; // Num vertical nodes.
+    let numk = 50; // Num horizontal nodes.
     let shift = 1.0;
     let i, k;
 
@@ -19,8 +17,7 @@ export function initWorld(RAPIER, testbed) {
                 .setTranslation(k * shift, -i * shift);
             let child = world.createRigidBody(bodyDesc);
             let colliderDesc = RAPIER.ColliderDesc.ball(rad);
-            let collider = world.createCollider(colliderDesc, child.handle);
-            let joint;
+            world.createCollider(colliderDesc, child.handle);
 
             // Vertical joint.
             if (i > 0) {
@@ -28,8 +25,7 @@ export function initWorld(RAPIER, testbed) {
                 let anchor1 = new RAPIER.Vector2(0.0, 0.0);
                 let anchor2 = new RAPIER.Vector2(0.0, shift);
                 let jointParams = RAPIER.JointParams.ball(anchor1, anchor2);
-                joint = world.createJoint(jointParams, parent, child);
-                joints.push(joint);
+                world.createJoint(jointParams, parent, child);
             }
 
             // Horizontal joint.
@@ -39,16 +35,14 @@ export function initWorld(RAPIER, testbed) {
                 let anchor1 = new RAPIER.Vector2(0.0, 0.0);
                 let anchor2 = new RAPIER.Vector2(-shift, 0.0);
                 let jointParams = RAPIER.JointParams.ball(anchor1, anchor2);
-                joint = world.createJoint(jointParams, parent, child);
-                joints.push(joint);
+                world.createJoint(jointParams, parent, child);
             }
 
             bodies.push(child);
-            colliders.push(collider);
         }
     }
 
-    testbed.setWorld(world, bodies, colliders, joints);
+    testbed.setWorld(world);
     testbed.lookAt({
         target: {x: 30.0, y: 30.0},
         zoom: 10.0